0.0.4
-ported woven armour, improved the properties
-ported bone tools, improved properties (TO DO: bone bow)
-edited tool tags to make bone axe work to mine wood
-changed woven armour to include brigadine in the name (a type of armour made of fabric or leather)
-other stuff idk i forgot

+++ src/main/java/net/teaper/aenarim/item/custom/ModHoeItem.java
@@ -4,7 +4,7 @@ import net.minecraft.item.HoeItem;
 import net.minecraft.item.ToolMaterial;
 
 public class ModHoeItem extends HoeItem {
-    public ModHoeItem(ToolMaterial material, int attackDamage, float attackSpeed, Settings settings) {
-        super(material, attackDamage, attackSpeed, settings);
+    public ModHoeItem(ToolMaterial material, float attackDamage, float attackSpeed, Settings settings) {
+        super(material, (int) attackDamage, attackSpeed, settings);
     }
 }
